Bluetooth: hci_sync: Make use of hci_cmd_sync_queue set 2
This make use of hci_cmd_sync_queue for the following MGMT commands:

Add Advertising
Remove Advertising
Add Extended Advertising Parameters
Add Extended Advertising Data

mgmt-tester -s "Add Advertising"

Test Summary
------------
Add Advertising - Failure: LE off                    Passed       0.175 seconds
Add Advertising - Invalid Params 1 (AD too long)     Passed       0.017 seconds
Add Advertising - Invalid Params 2 (Malformed len)   Passed       0.016 seconds
Add Advertising - Invalid Params 3 (Malformed len)   Passed       0.015 seconds
Add Advertising - Invalid Params 4 (Malformed len)   Passed       0.016 seconds
Add Advertising - Invalid Params 5 (AD too long)     Passed       0.020 seconds
Add Advertising - Invalid Params 6 (ScRsp too long)  Passed       0.020 seconds
Add Advertising - Invalid Params 7 (Malformed len)   Passed       0.020 seconds
Add Advertising - Invalid Params 8 (Malformed len)   Passed       0.021 seconds
Add Advertising - Invalid Params 9 (Malformed len)   Passed       0.021 seconds
Add Advertising - Invalid Params 10 (ScRsp too long) Passed       0.014 seconds
Add Advertising - Rejected (Timeout, !Powered)       Passed       0.022 seconds
Add Advertising - Success 1 (Powered, Add Adv Inst)  Passed       0.028 seconds
Add Advertising - Success 2 (!Powered, Add Adv Inst) Passed       0.027 seconds
Add Advertising - Success 3 (!Powered, Adv Enable)   Passed       0.019 seconds
Add Advertising - Success 4 (Set Adv on override)    Passed       0.019 seconds
Add Advertising - Success 5 (Set Adv off override)   Passed       0.019 seconds
Add Advertising - Success 6 (Scan Rsp Dta, Adv ok)   Passed       0.020 seconds
Add Advertising - Success 7 (Scan Rsp Dta, Scan ok)  Passed       0.020 seconds
Add Advertising - Success 8 (Connectable Flag)       Passed       0.019 seconds
Add Advertising - Success 9 (General Discov Flag)    Passed       0.018 seconds
Add Advertising - Success 10 (Limited Discov Flag)   Passed       0.018 seconds
Add Advertising - Success 11 (Managed Flags)         Passed       0.018 seconds
Add Advertising - Success 12 (TX Power Flag)         Passed       0.017 seconds
Add Advertising - Success 13 (ADV_SCAN_IND)          Passed       0.021 seconds
Add Advertising - Success 14 (ADV_NONCONN_IND)       Passed       0.018 seconds
Add Advertising - Success 15 (ADV_IND)               Passed       0.017 seconds
Add Advertising - Success 16 (Connectable -> on)     Passed       0.019 seconds
Add Advertising - Success 17 (Connectable -> off)    Passed       0.019 seconds
Add Advertising - Success 18 (Power -> off, Remove)  Passed       0.018 seconds
Add Advertising - Success 19 (Power -> off, Keep)    Passed       0.033 seconds
Add Advertising - Success 20 (Add Adv override)      Passed       0.019 seconds
Add Advertising - Success 21 (Timeout expires)       Passed       1.050 seconds
Add Advertising - Success 22 (LE -> off, Remove)     Passed       0.025 seconds
Add Advertising - Success (Empty ScRsp)              Passed       0.016 seconds
Add Advertising - Success (ScRsp only)               Passed       0.024 seconds
Add Advertising - Invalid Params (ScRsp too long)    Passed       0.016 seconds
Add Advertising - Success (ScRsp appear)             Passed       0.016 seconds
Add Advertising - Invalid Params (ScRsp appear long) Passed       0.017 seconds
Add Advertising - Success (Appear is null)           Passed       0.016 seconds
Add Advertising - Success (Name is null)             Passed       0.018 seconds
Add Advertising - Success (Complete name)            Passed       0.020 seconds
Add Advertising - Success (Shortened name)           Passed       0.019 seconds
Add Advertising - Success (Short name)               Passed       0.019 seconds
Add Advertising - Success (Name + data)              Passed       0.020 seconds
Add Advertising - Invalid Params (Name + data)       Passed       0.016 seconds
Add Advertising - Success (Name+data+appear)         Passed       0.022 seconds
Total: 47, Passed: 47 (100.0%), Failed: 0, Not Run: 0
Overall execution time: 2.17 seconds

mgmt-tester -s "Remove Advertising"

Test Summary
------------
Remove Advertising - Invalid Params 1                Passed       0.017 seconds
Remove Advertising - Success 1                       Passed       0.020 seconds
Remove Advertising - Success 2                       Passed       0.019 seconds
Total: 3, Passed: 3 (100.0%), Failed: 0, Not Run: 0
Overall execution time: 0.0585 seconds

mgmt-tester -s "Ext Adv MGMT Params"

Test Summary:
------------
Ext Adv MGMT Params - Unpowered                      Passed       0.019 seconds
Ext Adv MGMT Params - Invalid parameters             Passed       0.015 seconds
Ext Adv MGMT Params - Success                        Passed       0.016 seconds
Ext Adv MGMT Params - (5.0) Success                  Passed       0.019 seconds
Total: 4, Passed: 4 (100.0%), Failed: 0, Not Run: 0
Overall execution time: 0.0746 seconds

mgmt-tester -s "Ext Adv MGMT -"

Test Summary
------------
Ext Adv MGMT - Data set without Params               Passed       0.019 seconds
Ext Adv MGMT - AD Data (5.0) Invalid parameters      Passed       0.018 seconds
Ext Adv MGMT - AD Data (5.0) Success                 Passed       0.019 seconds
Ext Adv MGMT - AD Scan Response (5.0) Success        Passed       0.019 seconds
Total: 4, Passed: 4 (100.0%), Failed: 0, Not Run: 0
Overall execution time: 0.0805 seconds

+++ include/net/bluetooth/hci_core.h
@@ -1667,8 +1667,8 @@ struct sk_buff *__hci_cmd_sync(struct hci_dev *hdev, u16 opcode, u32 plen,
 			       const void *param, u32 timeout);
 struct sk_buff *__hci_cmd_sync_ev(struct hci_dev *hdev, u16 opcode, u32 plen,
 				  const void *param, u8 event, u32 timeout);
-int __hci_cmd_sync_status(struct hci_dev *hdev, u16 opcode, u32 plen,
-			  const void *param, u32 timeout);
+u8 __hci_cmd_sync_status(struct hci_dev *hdev, u16 opcode, u32 plen,
+			 const void *param, u32 timeout);
 int __hci_cmd_send(struct hci_dev *hdev, u16 opcode, u32 plen,
 		   const void *param);
